+++ plugins/form/classes/form.php
@@ -2,14 +2,36 @@
 namespace Grav\Plugin;
 
 use Grav\Common\Iterator;
-use Grav\Common\Grav;
 use Grav\Common\GravTrait;
 use Grav\Common\Page\Page;
+use Grav\Common\Data\Data;
+use Grav\Common\Data\Blueprint;
+use Grav\Common\Utils;
 use RocketTheme\Toolbox\Event\Event;
 
 class Form extends Iterator
 {
     use GravTrait;
+
+    /**
+     * @var string
+     */
+    public $message;
+
+    /**
+     * @var array
+     */
+    protected $data = array();
+
+    /**
+     * @var array
+     */
+    protected $rules = array();
+
+    /**
+     * @var array
+     */
+    protected $items = array();
 
     /**
      * @var array
@@ -39,6 +61,11 @@ class Form extends Iterator
         if (empty($this->items['name'])) {
             $this->items['name'] = $page->slug();
         }
+
+        $this->reset();
+
+        // Fire event
+        self::getGrav()->fireEvent('onFormInitialized', new Event(['form' => $this]));
     }
 
     /**
@@ -62,11 +89,11 @@ class Form extends Iterator
         if (!$name) {
             return $this->values;
         }
-        return $this->getField($name, 'values');
+        return $this->values->get($name);
     }
 
     /**
-     * Get value of given variable (or all values).
+     * Set value of given variable.
      *
      * @param string $name
      * @return mixed
@@ -77,7 +104,7 @@ class Form extends Iterator
             return;
         }
 
-        $this->values[$name] = $value;
+        $this->values->set($name, $value);
     }
 
     /**
@@ -85,7 +112,20 @@ class Form extends Iterator
      */
     public function reset()
     {
-        $this->values = array();
+        $name = $this->items['name'];
+
+        // Fix naming for fields (presently only for toplevel fields)
+        foreach ($this->items['fields'] as $key => $field) {
+            if (is_numeric($key) && isset($field['name'])) {
+                unset($this->items['fields'][$key]);
+
+                $key = $field['name'];
+                $this->items['fields'][$key] = $field;
+            }
+        }
+
+        $blueprint = new Blueprint($name, ['form' => $this->items, 'rules' => $this->rules]);
+        $this->values = new Data($this->data, $blueprint);
     }
 
     /**
@@ -94,16 +134,40 @@ class Form extends Iterator
     public function post()
     {
         if (isset($_POST)) {
-            $this->values = (array) $_POST;
+            $values = (array) $_POST;
+
+            if (method_exists('Grav\Common\Utils', 'getNonce')) {
+                if (!isset($values['form-nonce']) || !Utils::verifyNonce($values['form-nonce'], 'form')) {
+                    $event = new Event(['form' => $this, 'message' => self::getGrav()['language']->translate('PLUGIN_FORM.NONCE_NOT_VALIDATED')]);
+                    self::getGrav()->fireEvent('onFormValidationError', $event);
+                    return;
+                }
+            }
+
+            unset($values['form-nonce']);
+
+            foreach($this->items['fields'] as $field) {
+                if ($field['type'] == 'checkbox') {
+                    $name = $field['name'];
+                    $values[$name] = isset($values[$name]) ? true : false;
+                }
+            }
+
+            // Add post values to form dataset
+            $this->values->merge($values);
         }
 
-        foreach($this->items['fields'] as $field) {
-            if ($field['type'] == 'checkbox') {
-                if (isset($this->values[$field['name']])) {
-                    $this->values[$field['name']] = true;
-                } else {
-                    $this->values[$field['name']] = false;
-                }
+        // Validate and filter data
+        try {
+            $this->values->validate();
+            $this->values->filter();
+
+            self::getGrav()->fireEvent('onFormValidationProcessed', new Event(['form' => $this]));
+        } catch (\RuntimeException $e) {
+            $event = new Event(['form' => $this, 'message' => $e->getMessage()]);
+            self::getGrav()->fireEvent('onFormValidationError', $event);
+            if ($event->isPropagationStopped()) {
+                return;
             }
         }
 
@@ -119,30 +183,5 @@ class Form extends Iterator
         } else {
             // Default action.
         }
-    }
-
-
-    /**
-     * @param string $name
-     * @param string $scope
-     * @return mixed|null
-     * @internal
-     */
-    protected function getField($name, $scope = 'value')
-    {
-        $path = explode('.', $name);
-
-        $current = $this->{$scope};
-        foreach ($path as $field) {
-            if (is_object($current) && isset($current->{$field})) {
-                $current = $current->{$field};
-            } elseif (is_array($current) && isset($current[$field])) {
-                $current = $current[$field];
-            } else {
-                return null;
-            }
-        }
-
-        return $current;
     }
 }