+++ plugins/login/classes/ControllerInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Grav\Plugin\Login;
+
+interface ControllerInterface
+{
+    /**
+     * Performs an action.
+     */
+    public function execute();
+
+    /**
+     * Redirects an action
+     */
+    public function redirect();
+}