+++ plugins/login/login.php
@@ -1,26 +1,29 @@
 <?php
 namespace Grav\Plugin;
 
-use Grav\Common\Language\Language;
-use Grav\Common\Page\Page;
-use Grav\Common\Page\Pages;
-use Grav\Common\Plugin;
-use Grav\Common\Twig;
-use Grav\Common\Uri;
-use Grav\Common\User\User;
 use Grav\Plugin\Admin;
+use Grav\Common\Utils;
+use Grav\Common\Plugin;
+use Grav\Common\Page\Page;
+use Grav\Common\User\User;
+
 use RocketTheme\Toolbox\Session\Message;
-use RocketTheme\Toolbox\Session\Session;
 
 class LoginPlugin extends Plugin
 {
-    /** @var string */
+    /**
+     * @var string
+     */
     protected $route;
 
     /**
      * @var bool
      */
     protected $authenticated = true;
+
+    /**
+     * @var bool
+     */
     protected $authorized = true;
 
     /**
@@ -43,6 +46,19 @@ class LoginPlugin extends Plugin
      */
     public function initialize()
     {
+        /** @var Uri $uri */
+        $uri = $this->grav['uri'];
+
+        /** @var Grav\Common\Session */
+        $session = $this->grav['session'];
+
+        // Autoload classes
+        $autoload = __DIR__ . '/vendor/autoload.php';
+        if (!is_file($autoload)) {
+            throw new \Exception('Login Plugin failed to load. Composer dependencies not met.');
+        }
+        require_once $autoload;
+
         // Define session message service.
         $this->grav['messages'] = function ($c) {
             $session = $c['session'];
@@ -60,10 +76,44 @@ class LoginPlugin extends Plugin
 
             if (!isset($session->user)) {
                 $session->user = new User;
+
+                if ($c['config']->get('plugins.login.rememberme.enabled')) {
+                    $controller = new Login\Controller($this->grav, '');
+                    $rememberMe = $controller->rememberMe();
+
+                    // If we can present the correct tokens from the cookie, we are logged in
+                    $username = $rememberMe->login();
+                    if ($username) {
+                        // Normal login process
+                        $user = User::load($username);
+                        if ($user->exists()) {
+                            // There is a chance that an attacker has stolen
+                            // the login token, so we store the fact that
+                            // the user was logged in via RememberMe
+                            // (instead of login form)
+                            $session->remember_me = $rememberMe;
+                            $session->user = $user;
+                        }
+                    }
+
+                    // Check if the token was invalid
+                    if ($rememberMe->loginTokenWasInvalid()) {
+                        $controller->setMessage($t->translate('LOGIN_PLUGIN.REMEMBER_ME_STOLEN_COOKIE'));
+                    }
+                }
             }
 
             return $session->user;
         };
+
+        // Manage OAuth login
+        $task = !empty($_POST['task']) ? $_POST['task'] : $uri->param('task');
+        if (!$task && isset($_POST['oauth']) || (!empty($_GET) && $session->oauth)) {
+            $this->oauthController();
+        }
+
+        // Aborted OAuth authentication (invalidate it)
+        unset($session->oauth);
 
         // Register route to login page if it has been set.
         $this->route = $this->config->get('plugins.login.route');
@@ -74,6 +124,9 @@ class LoginPlugin extends Plugin
         }
     }
 
+    /**
+     * Add Login page
+     */
     public function addLoginPage()
     {
         /** @var Pages $pages */
@@ -90,6 +143,9 @@ class LoginPlugin extends Plugin
         }
     }
 
+    /**
+     * Initialize login controller
+     */
     public function loginController()
     {
         /** @var Uri $uri */
@@ -98,14 +154,52 @@ class LoginPlugin extends Plugin
         $task = substr($task, strlen('login.'));
         $post = !empty($_POST) ? $_POST : [];
 
-        require_once __DIR__ . '/classes/controller.php';
-        $controller = new LoginController($this->grav, $task, $post);
+        if (method_exists('Grav\Common\Utils', 'getNonce')) {
+            if ($task == 'login') {
+                if (!isset($post['login-form-nonce']) || !Utils::verifyNonce($post['login-form-nonce'], 'login-form')) {
+                    $this->grav['messages']->add($this->grav['language']->translate('LOGIN_PLUGIN.ACCESS_DENIED'), 'info');
+                    $this->authenticated = false;
+                    $twig = $this->grav['twig'];
+                    $twig->twig_vars['notAuthorized'] = true;
+                    return;
+                }
+            } else if ($task == 'logout') {
+                $nonce = $this->grav['uri']->param('logout-nonce');
+                if (!isset($nonce) || !Utils::verifyNonce($nonce, 'logout-form')) {
+                    return;
+                }
+            }
+        }
+
+        $controller = new Login\LoginController($this->grav, $task, $post);
         $controller->execute();
         $controller->redirect();
     }
 
+    /**
+     * Initialize OAuth login controller
+     */
+    public function oauthController()
+    {
+        /** @var Uri $uri */
+        $uri = $this->grav['uri'];
+        $oauth = !empty($_POST['oauth']) ? $_POST['oauth'] : $uri->param('oauth');
+        $oauth = $oauth ?: $this->grav['session']->oauth;
+        $post = !empty($_POST) ? $_POST : [];
+
+        $controller = new Login\OAuthLoginController($this->grav, $oauth, $post);
+        $controller->execute();
+        $controller->redirect();
+    }
+
+    /**
+     * Authorize Page
+     */
     public function authorizePage()
     {
+        /** @var User $user */
+        $user = $this->grav['user'];
+
         /** @var Page $page */
         $page = $this->grav['page'];
 
@@ -116,9 +210,6 @@ class LoginPlugin extends Plugin
         if (!$rules) {
             return;
         }
-
-        /** @var User $user */
-        $user = $this->grav['user'];
 
         // Continue to the page if user is authorized to access the page.
         foreach ($rules as $rule => $value) {
@@ -152,7 +243,6 @@ class LoginPlugin extends Plugin
             $twig = $this->grav['twig'];
             $twig->twig_vars['notAuthorized'] = true;
         }
-
     }
 
 
@@ -178,6 +268,17 @@ class LoginPlugin extends Plugin
 
         if (!$this->authenticated) {
             $twig->template = "login." . $extension . ".twig";
+
+            $providers = [];
+            foreach ($this->config->get('plugins.login.oauth.providers') as $provider => $options) {
+                if ($options['enabled'] && isset($options['credentials'])) {
+                    $providers[$provider] = $options['credentials'];
+                }
+            }
+            $twig->twig_vars['oauth'] = [
+                'enabled' => $this->config->get('plugins.login.oauth.enabled'),
+                'providers' => $providers
+            ];
         }
 
         // add CSS for frontend if required