+++ plugins/login/CHANGELOG.md
@@ -1,5 +1,17 @@
+# v1.0.0
+## 11/21/2015
+
+1. [](#new)
+    * Added OAuth login support for _Facebook_, _Google_, _GitHub_ and _Twitter_
+    * Added **Nonce** form security support
+    * Added option to "redirect after login"
+    * Added "remember me" functionality
+    * Added Hungarian translation
+2. [](#improved)
+    * Added blueprints for Grav Admin plugin (multi-language support!)
+
 # v0.3.3
-## 09/011/2015
+## 09/11/2015
 
 1. [](#improved)
     * Changed authorise to authorize
@@ -34,7 +46,7 @@
 
 1. [](#improved)
     * Disable `enable` in admin
-    
+
 # v0.1.0
 ## 08/04/2015